chore# user api 호출 변경
